Enforce simple polygon
Everytime the set of vertices is updated, it is sorted by
polar angle. Then the polygon is drawn based on the sorted set.
This ensures the polygon is always simple. This is the naive
approach to enforcing, makes creating complex shapes somewhat
difficult, but it gets the job done for what is needed in this
application.