Failed attempt to implement three coloring
Triangluate: Modify so that result of Process is a list of pointers
rather than copies, in attempt to make three coloring possible.
Everything here still works as it should.

SimplePolygon: Attempt to implement three coloring. It does not
currently work. The call to it is commented out, so app runs fine,
but no coloring of vertices.